Ignore the cluster network (team0)
OVS is used for 2 things:
- the VM network topology
- the cluster network (needed to enjoy RSTP, not available on standard
  linux bridges)

It's annoying that when changing the ovs VM topology, all bridges are
destroyed and created again. This creates connectivity problem for ceph
and corosync, can trigger fencing events, etc.

I propose to make this tool ignore the cluster network configuration
"team0", that we will create independantly with ansible.

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>