add port: allow the ofport_request option
When setting up openflow rules based on incoming port, we need to know the openflow port number ("ofport") value, which is decided at port creation time.
Since the rules will use a fixed ofport, we need to fix the ofport for the ports we need openflow on. OVS has the "ofport_request" option for this purpose.
This commit allows the seapath user to define the ofport_request value for a port in its OVS inventory, and it will be configured as requested at port creation.

Signed-off-by: Florent Carli <florent.carli@rte-france.com>